State observer is a new class. Move everything to UI. aurusov/raox#297
