+++ ru.bmstu.rk9.rao.ui/src/ru/bmstu/rk9/rao/ui/player/ModelStateStorage.java
@@ -1,20 +1,18 @@
-package ru.bmstu.rk9.rao.lib.simulator;
+package ru.bmstu.rk9.rao.ui.player;
 
 import java.util.ArrayList;
 import java.util.List;
+
+import ru.bmstu.rk9.rao.lib.simulator.ModelState;
 
 public class ModelStateStorage {
 	private List<ModelState> stateList = new ArrayList<ModelState>();
 	
 	void addModelState(ModelState modelState){
-		System.out.println("Adding modelState to list");
 		stateList.add(modelState);		
 	}
 	
 	List<ModelState> getStateList (){
-		
-		System.out.println("State list out");
 		return stateList;
-		
 	}
 }
