+++ ru.bmstu.rk9.rao.ui/src/ru/bmstu/rk9/rao/ui/simulation/RuntimeComponents.java
@@ -1,10 +1,12 @@
 package ru.bmstu.rk9.rao.ui.simulation;
 
 import ru.bmstu.rk9.rao.ui.notification.RealTimeUpdater;
+import ru.bmstu.rk9.rao.ui.player.StateObserver;
 
 public class RuntimeComponents {
 	public static RealTimeUpdater realTimeUpdater = null;
 	public static SimulationSynchronizer simulationSynchronizer = null;
+	public static StateObserver stateObserver = null;
 
 	private static boolean isInitialized = false;
 
@@ -15,6 +17,7 @@ public class RuntimeComponents {
 	public static final void initialize() {
 		realTimeUpdater = new RealTimeUpdater();
 		simulationSynchronizer = new SimulationSynchronizer();
+		stateObserver = new StateObserver();
 		isInitialized = true;
 	}
 
@@ -22,5 +25,7 @@ public class RuntimeComponents {
 		isInitialized = false;
 		realTimeUpdater.deinitialize();
 		simulationSynchronizer.deinitializeSubscribers();
+		stateObserver.deinitializeSubscribers();
+		
 	}
 }
